Extension tasks
After finishing the first task, the following other tasks should be
tackled:
* The input is extended by a second parameter `n` that restricts the
maximum length of the subsequence to n.
* The input is extended by a third parameter which changes how the
metric is calculated.
  Instead of the _sum_ of the values we want to find the highest _sum_
of the absolute values of the differences of neighboring pairs.
  As an example, for the input above the absolute differences would be:
`8 6 1 3 5 7 4 3` and therefore the output should be `16` for `n = 4` as
`-1 4 -3 1` result in the absolute differences of `5 7 4` which adds up
to `16`.
  The expected input parameter is `values` for the original behavior and
`differences` for the new one.